Make page contents wider
Main goal of these changes is to make page contents wider.

Official version of library has rather narrow (600px) space for
page contents. This is supposedly because it tries to mimic how
documents appear in Google Docs. However, in Reaktor Library we
want the contents to use more of the available horizontal space.

This required the following changes:

1) `layouts/pages/categories.ejs` had to be customized to add
new CSS div id `category-page-upper`

2) `layouts/partials/nav.ejs` had to be customized to add new
CSS class `g-upper-body-title`. Also, `partials/search.ejs` and
`partials/warning.ejs` had to be copied to custom folder, even
though they are identical to official versions. But if they are
not copied, the includes in `partials/nav.ejs` break.

3) The actual layout changes are in `styles/_custom.css`. The
gist is to give content a maximum of 900px horizontally instead
of the default 600px. The rest of the changes are there to
ensure that the table of contents (left panel) and the page
title (upper body) keep in line with the contents on all screen
widths.