Move "Pages in folder" to left panel in desktop layout
The "Pages in folder X" section that is by default shown at the
bottom of category index page has now been moved to left panel
below the table of contents, to improve its discoverability. This
change only affects the desktop layout. On mobile/tablet layout
the page listing is still shown at the bottom of the index page.

Layout file `categories/default.ejs` had to be customized for
this to happen. Also `partials/childrenList.ejs` was customized
to match the style of the page listing with the TOC title style.
Finally, a little bit of custom CSS was needed to make the items
in page listing match the visual representation of TOC items.

There is also one new div added in `categories/default.ejs` layout
file: `g-main-content-wrapper` is used in `public/scripts/main.js`
(unfortunately not customizable) to fix a bug that broke scrolling
of TOC/page listing in cases where the listing is long but the
actual contents of the page are very short (vertically).