Fix issue with Create New Page button and multiple Google accounts
The buttons "Create New Page" and "View Folder" did not work as
expected in case the logged in user has logged into Google with
multiple Google accounts. The problem is in the `/u/0` part of the
Google Docs and Google Drive URLs, which expects that the user has
logged in with his/her primary (first) Google account. But if the
user has a personal Google account as the primary one, and the
organization account (that grants access to Library) as the
secondary one, the links work incorrectly ("Create New Page"
creates a new document under user's personal account, and "View
Folder" shows a permission error).

With this change, the `/u/0` part is left out of the URLs, and
a query parameter `authuser` is added to the URLs. The `authuser`
is set to user's email, received from Google when the user logged
in. This way, the links are handled correctly (using the same
Google account that the user used to log in to Library).

Implementing this required a custom version of `footer.ejs`. The
only changes are the CSS ids that have been added to the buttons
and the code on lines 60-65 that rewrite the hrefs when the user
information has been retrieved from the backend.