Add some margin for list elements
List elements (`ul` and `ol`) in body text did not have any bottom
margin. This made the content look a bit cramped because the
following paragraph started right after the list.

This has been now fixed so that non-empty paragraphs that follow
a list have a top margin of 20px. It was better to implement it
this way, because in nested lists, it does not look good if the
different list levels have margin between them. And the paragraph
must be non-empty, since there is sometimes an empty paragraph
after a list, and that paragraph already has a bottom margin of
20px. And there is no need to double that.